Добавлены исправления

+++ main-service/src/main/java/ru/practicum/server/dto/eventDtos/EventFullDto.java
@@ -12,7 +12,7 @@ import java.time.LocalDateTime;
 public class EventFullDto {
     private String annotation;
     private CategoryDto category;
-    private Integer confirmedRequests;
+    private Long confirmedRequests;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss")
     private LocalDateTime createdOn;
     private String description;
@@ -27,5 +27,5 @@ public class EventFullDto {
     private Boolean requestModeration;
     private StateEnum state;
     private String title;
-    private Long views;
+    private Integer views;
 }
